Splitting up debt_payment into a package.
This helps to avoid a circular reference with TransactionStrategy.
Helper methods moved to util sub-module (util).
